A-Packages Divide classes into packages. Put all classes into duke package, organize commands in duke.command
