--- src/main/java/CommandNotFound.java
@@ -1,11 +1,0 @@
-public class CommandNotFound extends Command{
-
-    public CommandNotFound() {
-
-    }
-
-    @Override
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
-        ui.showCommandNotFound();
-    }
-}
