--- src/main/java/ListCommand.java
@@ -1,10 +1,0 @@
-public class ListCommand extends Command{
-
-    public ListCommand() {
-
-    }
-    @Override
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
-        tasks.print();
-    }
-}
