+++ src/main/java/duke/command/Command.java
@@ -1,0 +1,14 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public abstract class Command {
+
+    public abstract void execute(TaskList tasks, Ui ui, Storage storage);
+
+    public boolean isExit() {
+        return false;
+    }
+}
