+++ src/main/java/duke/command/ListCommand.java
@@ -1,0 +1,16 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
+public class ListCommand extends Command {
+
+    public ListCommand() {
+
+    }
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        tasks.print();
+    }
+}
