+++ src/main/java/duke/command/MarkCommand.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.Storage;
+import duke.TaskList;
+import duke.Ui;
+
 public class MarkCommand extends Command {
 
     private int index;
@@ -6,7 +12,7 @@ public class MarkCommand extends Command {
         this.index = index - 1;
     }
     @Override
-    public void execute(TaskList tasks,Ui ui,Storage storage) {
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
         tasks.mark(index);
         ui.showMarkSucess(tasks.get(index));
     }
